pid changes and yaw motor reset

+++ aruw-mcb-project/src/aruwsrc/control/turret/turret_motor.hpp
@@ -113,7 +113,7 @@ public:
      */
     mockable inline float getChassisFrameVelocity() const
     {
-        return (M_TWOPI / 60) * motor->getShaftRPM();
+        return (M_TWOPI / 60) * static_cast<float>(motor->getShaftRPM());
     }
 
     /**
